requirements.txt
# FastAPI e servidor
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Validação de dados
pydantic==2.10.6
pydantic-settings==2.7.1

# HTTP client
httpx==0.28.1

# OpenAI
openai==1.59.7

# Utilitários
python-dotenv==1.0.0